bpf: Fix host firewall when BPF NodePort is enabled
We want the host firewall to see ingress-to-host packets (1) after the
DNAT but before the SNAT on the way to the remote backend, and (2) after
the reverse SNAT but before the reverse DNAT for replies from remote
backend. This way, we always see packets with the backend pod's IP and the
client's IP.

See #12011 for details and discussion on this change.

Fixes: 88bf291 ("bpf: Enforce host policies for IPv4")
Fixes: 489dbef ("bpf: Enforce host policies for IPv6")
Related: https://github.com/cilium/cilium/pull/12011#discussion_r438624540
Signed-off-by: Paul Chaignon <paul@cilium.io>