bpf: Redirect local pod->host traffic through host fw
When traffic from a pod is destined to the its host, on egress from the
container, it is passed to the stack and doesn't go through the host
device (e.g., cilium_host). This results in a host firewall bypass on
ingress.

To fix this, we redirect traffic egressing pods to the host device when
the host firewall is enabled and the destination ID is that of the host.

Fixes: #11507
Signed-off-by: Paul Chaignon <paul@cilium.io>